pyproject.toml
[project]
name = "smsd"
version = "0.1.0"
description = ""
authors = [
    {name = "imsasankvindamuri",email = "imsasankvindamuri@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0.0"
dependencies = [
    "python-vlc (>=3.0.21203,<4.0.0)",
    "mutagen (>=1.47.0,<2.0.0)",
    
]

[tool.poetry]
packages = [{include = "smsd", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
smsd = "smsd.__main__:main"
